renamed index.js to app.js
