+++ lib/appFunctions.js
@@ -4,7 +4,7 @@ const inquirer = require('inquirer');
 const cTable = require('console.table');
 const prompts = require('./prompts');
 const queries = require('./queries');
-const db = require('../index');
+const app = require('../app');
 
 const roleChoices = [];
 const employeeChoices = [];
@@ -61,7 +61,7 @@ function prompt() {
           break;
 
         case prompts.exit:
-          db.connection.end();
+          app.connection.end();
           break;
         default:
           throw new Error('Unexpected prompt error.');
@@ -70,7 +70,7 @@ function prompt() {
 }
 
 function viewEmployees() {
-  db.connection.query(queries.allEmployees, (err, res) => {
+  app.connection.query(queries.allEmployees, (err, res) => {
     if (err) throw err;
     const employeeTable = cTable.getTable(res);
     console.log('\n' + employeeTable);
@@ -81,7 +81,7 @@ function viewEmployees() {
 function view(choice) {
   switch (choice) {
     case 'departments':
-      db.connection.query(queries.allDepartments, (err, res) => {
+      app.connection.query(queries.allDepartments, (err, res) => {
         if (err) throw err;
         const departmentTable = cTable.getTable(res);
         console.log('\n' + departmentTable);
@@ -89,7 +89,7 @@ function view(choice) {
       });
       break;
     case 'roles':
-      db.connection.query(queries.allRoles, (err, res) => {
+      app.connection.query(queries.allRoles, (err, res) => {
         if (err) throw err;
         const roleTable = cTable.getTable(res);
         console.log('\n' + roleTable);
@@ -153,7 +153,7 @@ async function addEmployee() {
       let managerIdIndex = employeeChoices.indexOf(answer.manager);
       let managerId = employeeIds[managerIdIndex];
 
-      db.connection.query(
+      app.connection.query(
         queries.addEmployee,
         {
           first_name: answer.firstName,
@@ -190,7 +190,7 @@ function addDepartment() {
       }
     ])
     .then(answer => {
-      db.connection.query(
+      app.connection.query(
         queries.addDepartment,
         {
           name: answer.departmentName
@@ -243,7 +243,7 @@ function addRole() {
     .then(answer => {
       let departmentId = departmentChoices.indexOf(answer.roleDepartment) + 1;
 
-      db.connection.query(
+      app.connection.query(
         queries.addRole,
         {
           title: answer.roleName,
@@ -280,7 +280,7 @@ function updateEmployeeRole() {
       let employeeId = employeeIds[employeeIdIndex];
       let newRole = roleChoices.indexOf(answer.roleUpdate) + 1;
 
-      db.connection.query(
+      app.connection.query(
         queries.updateEmployeeRole,
         [
           {
@@ -319,14 +319,14 @@ function getDatabaseInfo(table, column) {
   if (table === 'employee' && column === 'full_name') {
     employeeChoices.length = 0;
 
-    db.connection.query(queries.employeesFullName, (err, res) => {
+    app.connection.query(queries.employeesFullName, (err, res) => {
       if (err) throw err;
       for (const employees of res) {
         choiceArr.push(employees.full_name);
       }
     });
   } else {
-    db.connection.query(`SELECT ${column} FROM ${table}`, (err, res) => {
+    app.connection.query(`SELECT ${column} FROM ${table}`, (err, res) => {
       if (err) throw err;
 
       switch (column) {
